+++ modules/javafx.base/src/shims/java/com/sun/javafx/binding/LazyObjectBindingStub.java
@@ -25,7 +25,7 @@
 
 package com.sun.javafx.binding;
 
-import javafx.beans.Subscription;
+import javafx.util.Subscription;
 
 /**
  * Stub to allow testing of package private LazyObjectBinding.
